requirements.txt
Flask==2.0.1
SQLAlchemy==1.4.23
Flask-SQLAlchemy==2.5.1
Flask-Cors==3.0.10
PyJWT==2.1.0
gunicorn==20.1.0
psycopg2-binary==2.8.6
Werkzeug==2.0.3
pymysql==1.0.2
